% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_prcp.R
\docType{data}
\name{GeomRainfallRunoff}
\alias{GeomRainfallRunoff}
\title{`不知道如何传递给 scale_y_prcp 函数中`
`直接改变 data，或许可以避免使用coef`
`draw_panel 和 draw_group 的区别`}
\format{
An object of class \code{GeomRainfallRunoff} (inherits from \code{GeomLine}, \code{GeomPath}, \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomRainfallRunoff
}
\description{
`不知道如何传递给 scale_y_prcp 函数中`
`直接改变 data，或许可以避免使用coef`
`draw_panel 和 draw_group 的区别`
}
\keyword{datasets}
