% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runoff_data.R
\docType{data}
\name{runoff_data}
\alias{runoff_data}
\title{Runoff Data}
\format{
A data frame with variables:
  \describe{
    \item{time}{Time measurements}
    \item{Q}{Runoff measurements}
    \item{prcp}{Precipitation measurements}
  }
}
\usage{
runoff_data
}
\description{
This dataset contains runoff data for a specific location.
}
\examples{
# Print the first few rows of the runoff_data dataset
head(runoff_data)
}
\keyword{datasets}
